NODE-2066: Moving failed transactions tests to the new suite
